refactor config and git code into their own files
also handle walking paged responses for labels
